# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_gobravedev_gobrave_internal_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_gobravedev_gobrave_internal_errors.ErrorCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2100
    - 2101
    - 2102
    - 2103
    type: integer
    x-enum-varnames:
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrServiceUnavailable
    - ErrTimeout
    - ErrValidation
    - ErrTenantNotFound
    - ErrTenantAlreadyExists
    - ErrTenantInactive
    - ErrTenantNameRequired
    - ErrTenantInvalidStatus
    - ErrAgentMissingThinkingModel
    - ErrAgentMissingAllowedTools
    - ErrAgentInvalidMaxIterations
    - ErrAgentInvalidTemperature
...
  internal_handler.EditParamsV2Response:
    properties:
      analysis_id:
        type: string
      analysis_name:
        type: string
      analysis_result:
        additionalProperties: true
        type: object
      formJson:
        items: {}
        type: array
      is_cache:
        type: boolean
      is_report:
        type: boolean
      request_param:
        additionalProperties: true
        type: object
      server_status:
        type: string
      status:
        type: string
    type: object
  internal_handler.WorkflowFormJSONResponse:
    properties:
      formJson:
        items: {}
        type: array
      type:
        type: string
    type: object
  internal_handler.WorkflowFormResponse:
    properties:
      analysis_result:
        additionalProperties: true
        type: object
      formJson:
        items: {}
        type: array
      type:
        type: string
    type: object
...
info:
  contact:
    name: AiAgentGo Github
    url: https://github.com/minebiome/ai-agent-go
  description: AiAgentGo 是一个基于 Go 语言构建的 AI Agent 框架，旨在提供高性能、可扩展的解决方案，帮助开发者快速构建智能代理应用。
  termsOfService: http://swagger.io/terms/
  title: AiAgentGo API
  version: "1.0"
paths:
  /analysis/edit-params-v2/{analysisId}:
    post:
      description: 查询 analysis 基础信息，并按 workflowId 复用工作流表单逻辑返回 formJson 与 analysis_result
      parameters:
      - description: 分析 ID
        in: path
        name: analysisId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.EditParamsV2Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 编辑分析参数（V2）
      tags:
      - 分析
...
  /workflow/tools/get-from-json/{workflowId}:
    get:
      description: 根据 workflowId 解析工作流 DAG，聚合输入/参数/formJson 配置
      parameters:
      - description: 工作流 ID
        in: path
        name: workflowId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.WorkflowFormJSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 获取工作流表单配置
      tags:
      - 工作流
  /workflows/{workflowId}/form:
    get:
      description: 基于 workflowId 返回 formJson，并按 input_type 自动补充 analysis_result（sample
        与按 role 分组的文件）
      parameters:
      - description: 工作流 ID
        in: path
        name: workflowId
        required: true
        type: string
      - description: 项目业务ID
        in: query
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.WorkflowFormResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gobravedev_gobrave_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 获取工作流表单与分析数据
      tags:
      - 工作流
securityDefinitions:
  ApiKeyAuth:
    description: 租户身份认证：输入 sk- 开头的 API Key
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: 用户登录认证：输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"